feat: add the cumulative type to the tooltip if any
