+++ src/stories/containers/Finances/components/SectionPages/MakerDAOExpenseMetrics/MakerDAOChartMetrics/MakerDAOChartMetrics.tsx
@@ -1,13 +1,19 @@
 import styled from '@emotion/styled';
 import { useMediaQuery } from '@mui/material';
-import { createChartTooltip } from '@ses/containers/Finances/utils/chartTooltip';
-import { formatterBreakdownChart, getChartAxisLabelByGranularity } from '@ses/containers/Finances/utils/utils';
+import {
+  formatBudgetName,
+  formatterBreakdownChart,
+  getChartAxisLabelByGranularity,
+  removeBudgetWord,
+} from '@ses/containers/Finances/utils/utils';
 import { useThemeContext } from '@ses/core/context/ThemeContext';
-import { replaceAllNumberLetOneBeforeDot } from '@ses/core/utils/string';
+import { zIndexEnum } from '@ses/core/enums/zIndexEnum';
+import { formatNumber, replaceAllNumberLetOneBeforeDot } from '@ses/core/utils/string';
 import lightTheme from '@ses/styles/theme/light';
 import ReactECharts from 'echarts-for-react';
 import React, { useRef } from 'react';
-import type { LineChartSeriesData } from '@ses/containers/Finances/utils/types';
+import type { CumulativeType } from '../useMakerDAOExpenseMetrics';
+import type { BarChartSeries, LineChartSeriesData } from '@ses/containers/Finances/utils/types';
 import type { AnalyticGranularity } from '@ses/core/models/interfaces/analytic';
 import type { WithIsLight } from '@ses/core/utils/typesHelpers';
 import type { EChartsOption } from 'echarts-for-react';
@@ -16,6 +22,8 @@ interface BreakdownChartProps {
   selectedGranularity: AnalyticGranularity;
   series: LineChartSeriesData[];
   handleToggleSeries: (series: string) => void;
+  isCumulative: boolean;
+  cumulativeType: CumulativeType;
 }
 
 const MakerDAOChartMetrics: React.FC<BreakdownChartProps> = ({
@@ -23,6 +31,8 @@ const MakerDAOChartMetrics: React.FC<BreakdownChartProps> = ({
   selectedGranularity,
   series,
   handleToggleSeries,
+  isCumulative,
+  cumulativeType,
 }) => {
   const { isLight } = useThemeContext();
   const chartRef = useRef<EChartsOption | null>(null);
@@ -46,7 +56,113 @@ const MakerDAOChartMetrics: React.FC<BreakdownChartProps> = ({
   };
 
   const options: EChartsOption = {
-    tooltip: createChartTooltip(selectedGranularity, year, isLight, isMobile, false, false, false),
+    tooltip: {
+      show: !isMobile,
+      trigger: 'axis',
+      extraCssText: `z-index:${zIndexEnum.ECHART_TOOL_TIP}`,
+      axisPointer: {
+        type: 'shadow',
+        shadowStyle: {
+          color: isLight ? '#D4D9E1' : '#231536',
+          opacity: 0.15,
+        },
+      },
+      padding: 0,
+      borderWidth: 1,
+      position: function (
+        point: [number, number],
+        params: EChartsOption,
+        dom: EChartsOption,
+        rect: EChartsOption,
+        size: EChartsOption
+      ) {
+        const MORE_WITH = 10;
+        const withTooltip = size.contentSize[0];
+        const heightTooltip = size.contentSize[0];
+
+        let xPos = point[0];
+        let yPos = point[1];
+
+        const tooltipWidth = withTooltip;
+        const tooltipHeight = heightTooltip;
+
+        if (xPos + tooltipWidth + MORE_WITH > window.innerWidth) {
+          xPos -= tooltipWidth;
+        }
+
+        if (yPos + tooltipHeight + MORE_WITH > window.innerHeight) {
+          yPos -= tooltipHeight;
+        }
+      },
+      borderColor: isLight ? '#D4D9E1' : '#231536',
+      formatter: function (params: BarChartSeries[]) {
+        // If all values are cero, don't show tooltip
+        if (params.every((item) => item.value === 0)) {
+          return '';
+        }
+        const filteredParams =
+          // !isShowZeroValues
+          //   ?
+          params;
+        // : params.filter((item) => item.value !== 0 && Math.abs(item.value) > 0.004);
+
+        const shortAmount = params.length > 10;
+        const flexDirection = shortAmount ? 'row' : 'column';
+        const wrap = shortAmount ? 'flex-wrap:wrap;' : '';
+        const gap = shortAmount ? '16px' : '12px';
+        const minMax = isTablet
+          ? 'max-width:300px'
+          : isDesktop1024
+          ? 'max-width:400px'
+          : 'min-width:190px;max-width:450px';
+        const maxWithTable = isTablet ? 'max-width:190px' : isDesktop1024 ? 'max-width:450px' : '';
+
+        return `
+          <div style="background-color:${isLight ? '#fff' : '#000A13'};padding:16px;overflow:auto;border-radius:3px;">
+            <div style="display: flex;justify-content: space-between;gap:24px;text-align:center;">
+              <div style="margin-bottom:16px;font-size:12px;font-weight:600;color:#B6BCC2;">${
+                (selectedGranularity as string) === 'Annually' ? year : params?.[0]?.name
+              }</div>
+              ${
+                isCumulative
+                  ? `<div style="text-transform:uppercase;font-weight:300;font-size:11px;color:${
+                      isLight ? '#434358' : '#546978'
+                    }">${cumulativeType} cumulative</div>`
+                  : ''
+              }
+            </div>
+            <div style="display:flex;flex-direction:${flexDirection};gap:${gap};${wrap}${minMax}">
+              ${filteredParams
+                .reverse()
+                .map(
+                  (item) =>
+                    `<div style="display: flex;align-items:center;gap: 6px;">
+                  <svg xmlns="http://www.w3.org/2000/svg" width="${isMobile ? 13 : 16}" height="${
+                      isMobile ? 13 : 16
+                    }" viewBox="0 0 13 13" fill="none" style="min-width:${isMobile ? 13 : 16};min-height:${
+                      isMobile ? 13 : 16
+                    }">
+                    <circle cx="6.5" cy="6.5" r="5.5" stroke="${item.color}" />
+                    <circle cx="6.5" cy="6.5" r="4" fill="${item.color}" />
+                  </svg>
+                  <span style="display: inline-block;font-size:14px;color:${
+                    isLight ? '#231536' : '#B6BCC2'
+                  };white-space:nowrap;overflow:hidden;text-overflow:ellipsis;${maxWithTable}"> ${
+                      !isMobile
+                        ? formatBudgetName(item.seriesName)
+                        : removeBudgetWord(formatBudgetName(item.seriesName))
+                    }:</span>
+                  <span style="font-size:16px;font-weight:700;color:${
+                    isLight ? '#231536' : '#EDEFFF'
+                  };display: inline-block;">${formatNumber(item.value)}</span>
+                </div>`
+                )
+                .join('')}
+            </div>
+          </div>
+          `;
+      },
+    },
     grid: {
       height: isLessMobile ? 192 : isMobile ? 192 : isTablet ? 409 : isDesktop1024 ? 398 : isDesktop1280 ? 399 : 399,
       width: isLessMobile ? 290 : isMobile ? 304 : isTablet ? 645 : isDesktop1024 ? 704 : isDesktop1280 ? 955 : 955,
